Refactor code for consistency and clarity; Fix xUnit tests

+++ AppCore/Dto/PagedResult.cs
@@ -7,7 +7,7 @@ public record PagedResult<T>(
     int PageSize
 )
 {
-    public int TotalPages => (int)Math.Ceiling((double)TotalCount / (double)PageSize);
+    public int TotalPages => (int)Math.Ceiling(TotalCount / (double)PageSize);
     public bool HasNext => Page < TotalPages;
     public bool HasPrevious => Page > 1;
 }
